+++ LibSWBF2/LoggerEntry.cpp
@@ -6,14 +6,14 @@ namespace LibSWBF2::Logging
 	LoggerEntry::LoggerEntry(const char* message, const ELogType& level, const uint64_t& line, const char* file)
 	{
 		this->m_Message = message;
-		this->m_Level = level;
+		this->m_Level = (uint8_t)level;
 		this->m_Line = line;
 		this->m_File = file;
 	}
 
 	string LoggerEntry::ToString() const
 	{
-		if (m_Level == ELogType::Info)
+		if (m_Level == (uint8_t)ELogType::Info)
 		{
 			return "[" + SLogType[m_Level] + "] " + m_Message;
 		}
